merge_tools: move symbols to `external.rs`
